requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
docopt==0.6.2             # via -r requirements.in
numpy==1.17.2             # via -r requirements.in, opencv-python, scipy, torch
opencv-python==4.4.0.42   # via -r requirements.in
pillow==6.1.0             # via -r requirements.in
scipy==1.5.2              # via -r requirements.in
six==1.12.0               # via -r requirements.in
# torch==1.2.0              # via -r requirements.in
tqdm==4.49.0              # via -r requirements.in
